backend1/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
tensorflow
pydantic
requests
scikit-learn
